backend/requirements.txt
pymongo[srv]==3.12
fastapi
uvicorn
python-dotenv
openai
langchain
langgraph
pydantic
typing-extensions